Create maxSub.c
You are given three integers n, m, and k. In a single move, you can subtract any integer in the range from 1 to m from n, but you are not be allowed to subtract this integer during the next k moves. What is the minimum number of moves required to make n strictly smaller than 0?
 
Input format: The first and only line of input contains three integers - n,m,k 
Output format: Print an integer denoting the answer to the problem.